prototype_x1000/agents/pyproject.toml
[project]
name = "prototype-x1000-agents"
version = "0.1.0"
description = "Multi-agent framework for Prototype X1000 powered by Anthropic SDK"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "DropFly", email = "dev@dropfly.io" }
]
keywords = ["agents", "anthropic", "llm", "ai", "automation"]

dependencies = [
    "anthropic>=0.40.0",
    "supabase>=2.0.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
]
mcp = [
    "mcp>=1.0.0",
]

[project.scripts]
px1000 = "cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "ceo", "specialists", "brain_builder", "memory", "tools", "cli"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["core", "ceo", "specialists", "brain_builder", "memory", "tools", "cli"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
